Provide a stable version that avoids circular import between 'pjml' and 'pjautoml'.
It is not possible to separate the 'pjml' from the 'pjautoml' directly.
I made this 'commit' to leave a stable version, where there is no cyclical dependency.
I understand that everything that refers to configuration space and its manipulation belongs conceptually to 'pjautoml'.
That is the difference between working in ML and AutoML level.
I worked a lot creating other versions and I think I found the conceptual problem now.
I know how to propose something more interesting and I will do this in a few days :)